adds gem framework
