+++ Gemfile
@@ -1,0 +1,11 @@
+source :rubygems
+
+gem 'vcr'
+
+group :development, :test do
+  gem 'guard',      '>= 1.6.2'
+  gem 'pry',        '>= 0.9.10'
+  gem 'pry-doc',    '>= 0.4.4'
+  gem 'rake',       '>= 10.0.4'
+  gem 'rspec',      '>= 2.13'
+end
