+++ Rakefile
@@ -1,0 +1,2 @@
+require "bundler"
+Bundler::GemHelper.install_tasks
