+++ lib/vcr_proxy.rb
@@ -1,0 +1,9 @@
+# stdlib modules
+require 'net/http'
+require 'webrick'
+require 'webrick/https'
+require 'webrick/httpproxy'
+
+require 'vcr'
+
+require 'vcr_proxy/vcr_proxy'
