+++ lib/vcr_proxy/version.rb
@@ -1,0 +1,3 @@
+module VCRProxy
+  VERSION = '0.0.1'
+end
