+++ spec/vcr_proxy_spec.rb
@@ -1,0 +1,7 @@
+require 'spec_helper'
+
+describe VCRProxy do
+  it "should work" do
+    VCRProxy.start({}).should be_kind_of?(VCRProxy)
+  end
+end
